gossip: pr1 reflects pr2
