+++ src/protocol/gossip/impl/CMakeLists.txt
@@ -3,15 +3,16 @@
 
 libp2p_add_library(p2p_gossip
     common.cpp
-    gossip_wire_protocol.cpp
+    message_builder.cpp
+    message_parser.cpp
     peers.cpp
     message_cache.cpp
     )
 target_link_libraries(p2p_gossip
     Boost::boost
     p2p_byteutil
+    p2p_protocol_common
     p2p_peer_id
     p2p_cid
     p2p_gossip_proto
-    p2p_logger
     )
